Complete `calloc` implementation
GCC's static analyzer reports a `malloc` pointer leak, but I don't
believe that this is correct. If I had to guess, the fact that we're writing
our own implementation of `malloc` is confusing the analyzer. For future
reference, disable the `-Wno-analyzer-malloc-leak` flag in the kernel
SConscript to see the exact analyzer error